requirements.txt
Flask==2.2.5
pyngrok==7.1.6
python-dotenv==0.21.0
scikit-learn==1.2.2
TensorFlow==2.16.1
Keras==3.3.3
yfinance==0.2.40
NumPy==1.26.4
Pandas==2.1.4
Flask-RESTful== 0.3.10